tests/wpt/tests/svg/linking/scripted/a.type-getter-02.svg

<?xml version="1.0" encoding="UTF-8"?>
<svg xmlns="http://www.w3.org/2000/svg"
     xmlns:h="http://www.w3.org/1999/xhtml">
  <title>SVGAElement.type getter</title>
  <metadata>
    <h:link rel="help" href="https://svgwg.org/svg2-draft/linking.html#InterfaceSVGAElement"/>
  </metadata>
  <a id="test" href="a" type="text/plain"></a>
  <h:script src="/resources/testharness.js"/>
  <h:script src="/resources/testharnessreport.js"/>
  <script><![CDATA[
    test(function() {
      var a = document.getElementById("test");

      test(function() {
        assert_equals(a.type, "text/plain");
       }, "Test anchor's type getter");
    });
  ]]></script>
</svg>
