third_party/blink/web_tests/external/wpt/svg/linking/scripted/a.type-getter-02-expected.txt

This is a testharness.js-based test.
[FAIL] Test anchor's type getter
  assert_equals: expected (string) "text/plain" but got (undefined) undefined
Harness: the test ran to completion.
